sdk/python/examples/cli_chat/requirements.txt
permem
openai>=1.0.0
python-dotenv>=1.0.0
